requirements.txt
streamlit==1.31.1
llama-cpp-python==0.2.23
Pillow==10.2.0
requests==2.31.0 